feat: Search Page now up and running
Search bar is a controlled component that uses its state to hit the Books API to retrieve books and display it on the search page. Slight issue where sometimes the network requests are too slow relative to React being able to render and set state which can lead to unintended behavior at times. Can fix it in future commits by implementing a lag before React renders so that it does not keep firing off AJAX requests for every single alphabet a user types