feat: Added in React router and added in BrowserRouter, Link and Route components
There was a bug that took me awhile to debug. I thought it was an issue with React Router but it was something as simple as using {} brackets instead of () in the return statement after the Route component declaration. Also removed all redundant code from toggleSearchPage which is no longer needed now that I am rendering based on matching paths not change of state